+++ src/app/service/poap/EndPOAP.ts
@@ -21,7 +21,12 @@ export default async (guildMember: GuildMember, platform: string, ctx?: CommandC
 	Log.debug('authorized to end poap event');
 	
 	if (platform == constants.PLATFORM_TYPE_TWITTER) {
-		await endTwitterPOAPFlow(guildMember, db, ctx);
+		// await endTwitterPOAPFlow(guildMember, db, ctx);
+		if (ctx) {
+			await ctx.send('Coming soon...');
+		} else {
+			await guildMember.send({ content: 'Coming soon...' });
+		}
 		return;
 	}